+++ page.php
@@ -51,4 +51,4 @@ get_header(); ?>
 
 <?php endwhile; ?>
 
-<?php get_footer(); ?>
+<?php get_footer();