+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2016-03-07
+### Added
+- Images source and licence informations.
+
+### Changed
+- Removed web font loader script. Fonts are imported trough CSS.
+- Header Customizer section. Header controls are moved to default Site Identity section.
+- Front Page Customizer section name is replaced with default "Static Front Page".
+- Footer text is translatable.
+- Link to Background-check licence.
+
+### Removed
+- Custom background feature.
+- Blank line on the end of page.php.
+
 ## [1.0.1] - 2016-03-04
 ### Added
 - 'pine_portfolio_post_type', 'pine_portfolio_posts_per_page' and 'pine_portfolio_terms' filters